Custom Ohai plugin for installed emacs version
Create a custom Ohai plugin that provides the current emacs version
installed on the system. This value is used to check if the `compile`
recipe should be included or not instead of checking that build
directory exists.

This cookbook now requires the `ohai` cookbook as dependency to load the
custom `emacs` plugin located in `files/default/ohai_plugins/emacs.rb`.